+++ Paddle.cpp
@@ -10,7 +10,7 @@ Paddle::Paddle(int16_t x, int16_t y)
 : w(2), h(14), x(x), y(y), d(0)
 {}
 
-void Paddle::update(int16_t target_y, int16_t ball_dx, bool should_miss) {
+void Paddle::update(int16_t target_y, int8_t ball_dx, bool should_miss) {
   bool my_direction = false;
   // Check side
   if (x == 0) {
