+++ PongGame.h
@@ -9,14 +9,14 @@ class PongGame {
 public:
   PongGame();
 
-  void setScore(uint16_t hour, uint16_t minute);
-  void update(uint16_t hour, uint16_t minute);
+  void setScore(uint8_t hour, uint8_t minute);
+  void update(uint8_t hour, uint8_t minute);
   void draw(Adafruit_GFX* display) const;
   void drawScore(Adafruit_GFX* display) const;
 private:
   Paddle lpaddle, rpaddle;
   Ball ball;
-  int16_t lscore, rscore;
+  uint8_t lscore, rscore;
   uint8_t pause;
 };
